Fixed issue when delay stage software missing
The code previously crashed when the file Newport.DLS.CommandInterfaceDLS was not present. Modifications made such that if this file is not present, a warning message is printed to inform the user that the delay stage control software is not present, and the connect button on the delay stage tab is disabled.